+++ ci/cpu/build.sh
@@ -45,6 +45,10 @@ gcc --version
 g++ --version
 conda list
 
+# Possible "hack"/workaround for Gunrock submodule `--remote` issues (TODO: FIX ME!)
+#
+git submodule update --init --recursive thirdparty/gunrock/
+
 # FIX Added to deal with Anancoda SSL verification issues during conda builds
 conda config --set ssl_verify False